backend/requirements.txt
fastapi
uvicorn
python-jose
google-api-python-client
google-auth
google-auth-oauthlib
pydantic